cqlite_core::parser::enhanced_statistics_parser

Function parse_nb_format_header

pub fn parse_nb_format_header(input: &[u8]) -> IResult<&[u8], StatisticsHeader>

Enhanced Statistics.db header parser for real ‘nb’ format

This function parses the actual 32-byte binary header structure from Cassandra 5.0 Statistics.db files. Based on hex analysis of real files:

00000000  00 00 00 04 26 29 1b 05  00 00 00 00 00 00 00 2c
00000010  00 00 00 01 00 00 00 65  00 00 00 02 00 00 14 d4

Binary Format (32 bytes)

Bytes 0-3: version_type (u32 BE) - Format version identifier (e.g., 0x00000004)

Bytes 4-7: statistics_kind (u32 BE) - Statistics type marker (e.g., 0x26291b05)

Bytes 8-11: reserved1 (u32 BE) - Reserved field (typically 0x00000000)

Bytes 12-15: data_length (u32 BE) - Length of variable-length data section

Bytes 16-19: metadata1 (u32 BE) - Metadata field (purpose TBD in M2)

Bytes 20-23: metadata2 (u32 BE) - Metadata field (purpose TBD in M2)

Bytes 24-27: metadata3 (u32 BE) - Metadata field (purpose TBD in M2)

Bytes 28-31: checksum_or_more (u32 BE) - Checksum or additional metadata

Returns

Ok((remaining_input, StatisticsHeader)) on successful parse of 32-byte header.

Note

This is the ONLY function in this module that reads actual binary data. All other parsing functions have been removed per Issue #28 mandate.