cqlite_core::storage::sstable

Module directory

Re-exports

pub use types::SSTableComponent;

pub use types::SSTableGeneration;

pub use types::SecondaryIndex;

pub use validation::ComponentAnalysis;

pub use validation::ValidationReport;

pub use validation::test_all_directories;

pub use validation::test_directory_validation;

pub use validation::validate_generation_components;

pub use validation::validate_toc_consistency;

pub use validation::validate_generation_components_enhanced;

pub use validation::validate_toc_consistency_enhanced;

Modules

types

validation

Structs

SSTableDirectory

Represents an entire SSTable directory containing multiple generations

Functions

parse_toc_file

Parse TOC.txt file to get list of components with enhanced validation (backward compatibility)

parse_toc_file_detailed

Parse TOC.txt file to get list of components with enhanced validation Returns (valid_components, unknown_components)