cqlite_core::types

Enum Value

pub enum Value {

    Null,
    Boolean(bool),
    Integer(i32),
    BigInt(i64),
    Counter(i64),
    Float(f64),
    Text(String),
    Blob(Vec<u8>),
    Timestamp(i64),
    Date(i32),
    Time(i64),
    Uuid([u8; 16]),
    Varint(Vec<u8>),
    Decimal {
        scale: i32,
        unscaled: Vec<u8>,
    },
    Duration {
        months: i32,
        days: i32,
        nanos: i64,
    },
    Json(Value),
    TinyInt(i8),
    SmallInt(i16),
    Float32(f32),
    List(Vec<Value>),
    Set(Vec<Value>),
    Map(Vec<(Value, Value)>),
    Tuple(Vec<Value>),
    Udt(UdtValue),
    Frozen(Box<Value>),
    Tombstone(TombstoneInfo),
    Inet(Vec<u8>),

}

Database value type that can hold any supported data type

Variants

Null

Null value

Boolean(bool)

Boolean value

Integer(i32)

32-bit signed integer

BigInt(i64)

64-bit signed integer

Counter(i64)

Cassandra Counter type (distributed increment-only counter with CRDT semantics)

Float(f64)

64-bit floating point number

Text(String)

UTF-8 string

Blob(Vec<u8>)

Binary data

Timestamp(i64)

Timestamp (milliseconds since Unix epoch)

Date(i32)

Date (days since Unix epoch: 1970-01-01)

Time(i64)

Time (nanoseconds since midnight)

Uuid([u8; 16])

UUID as 16 bytes

Varint(Vec<u8>)

Variable-length integer

Decimal

Decimal value with scale and unscaled value

Fields

scale: i32

unscaled: Vec<u8>

Duration

Duration value with months, days, and nanoseconds

Fields

months: i32

days: i32

nanos: i64

Json(Value)

JSON value

TinyInt(i8)

8-bit signed integer (for exact Cassandra compatibility)

SmallInt(i16)

16-bit signed integer (for exact Cassandra compatibility)

Float32(f32)

32-bit floating point (for exact Cassandra compatibility)

List(Vec<Value>)

List of values

Set(Vec<Value>)

Set of values (implemented as Vec for ordering preservation)

Map(Vec<(Value, Value)>)

Map of key-value pairs (Vec of tuples for exact Cassandra format)

Tuple(Vec<Value>)

Tuple with fixed-size heterogeneous types

Udt(UdtValue)

User defined type with structured fields

Frozen(Box<Value>)

Frozen wrapper for collections (immutable)

Tombstone(TombstoneInfo)

Tombstone marker indicating deleted data

Inet(Vec<u8>)

IP address (4 bytes for IPv4, 16 bytes for IPv6)

Implementations

impl Value

pub fn data_type(&self) -> CqlType

Get the data type of this value

pub fn is_null(&self) -> bool

Check if this value is null

pub fn is_tombstone(&self) -> bool

Check if this value is a tombstone (deleted)

pub fn is_deleted(&self) -> bool

Check if this value is deleted (null or tombstone)

pub fn is_expired(&self, current_time: i64) -> bool

Check if this value has expired based on TTL

pub fn deletion_time(&self) -> Option<i64>

Get the deletion timestamp if this is a tombstone

pub fn row_tombstone(deletion_time: i64) -> Self

Create a row tombstone with the given timestamp

pub fn cell_tombstone(deletion_time: i64) -> Self

Create a cell tombstone with the given timestamp

pub fn ttl_tombstone(deletion_time: i64, ttl: i64) -> Self

Create a TTL expiration tombstone

pub fn range_tombstone( deletion_time: i64, start_key: RowKey, end_key: RowKey, ) -> Self

Create a range tombstone for clustering key ranges

pub fn range_tombstone_with_ttl( deletion_time: i64, start_key: RowKey, end_key: RowKey, ttl: i64, ) -> Self

Create a range tombstone with TTL for clustering key ranges

pub fn tombstone_type(&self) -> Option<TombstoneType>

Get the tombstone type if this is a tombstone

pub fn tombstone_covers_key(&self, key: &RowKey) -> bool

Check if this tombstone covers a specific key (for range tombstones)

pub fn tombstone_ttl(&self) -> Option<i64>

Get TTL information from tombstone

pub fn is_tombstone_type(&self, tombstone_type: TombstoneType) -> bool

Check if this is a specific type of tombstone

pub fn tombstone_range(&self) -> Option<(Option<&RowKey>, Option<&RowKey>)>

Get range information for range tombstones

pub fn as_bool(&self) -> Option<bool>

Try to convert this value to a boolean

pub fn as_i32(&self) -> Option<i32>

Try to convert this value to an integer

pub fn as_i64(&self) -> Option<i64>

Try to convert this value to a big integer

pub fn as_date(&self) -> Option<i32>

Try to convert this value to a date (days since epoch)

pub fn as_time(&self) -> Option<i64>

Try to convert this value to a time (nanoseconds since midnight)

pub fn as_f64(&self) -> Option<f64>

Try to convert this value to a float

pub fn as_str(&self) -> Option<&str>

Try to convert this value to a string

pub fn as_bytes(&self) -> Option<&[u8]>

Try to convert this value to bytes

pub fn as_inet_bytes(&self) -> Option<&[u8]>

Try to convert this value to IP address bytes

pub fn len(&self) -> usize

Get the length of the byte representation for this value Returns the length of the underlying bytes for Blob/Text, or 0 for other types

pub fn is_empty(&self) -> bool

Check if the value is empty (for Blob/Text) or is Null

pub fn size_estimate(&self) -> usize

Get the size in bytes for this value when serialized

pub fn is_empty_collection(&self) -> bool

Check if this value represents an empty collection

pub fn collection_len(&self) -> Option<usize>

Get the element count for collections

pub fn is_valid_collection_element(&self) -> bool

Check if this value can be used as a collection element

pub fn validate_collection_types(&self) -> Result<()>

Validate collection type consistency